move patch file, add comments
