+++ src/llmcompressor/modeling/patch/qwen3_omni_patch.py
